requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
openai
httpx
apscheduler
pydantic
langchain-openai
langchain-core